Add files via upload
The path visualizer project is a simple graphical application that allows users to draw a grid and visualize the A* search algorithm in action to find the shortest path between two points on the grid. Here are some of its key features:

Grid creation: The application allows users to create a grid of any size by specifying the number of rows and columns.

Start and end points: Users can designate a starting point and an ending point on the grid by clicking on the corresponding cells. These cells are colored green and red, respectively.

Walls: Users can designate cells on the grid as walls by clicking on them. These cells are colored black and cannot be traversed by the A* search algorithm.

Pathfinding: When the user presses the space bar, the application runs the A* search algorithm to find the shortest path between the start and end points. The algorithm takes into account the walls on the grid and generates a path that avoids them.

Path visualization: After the A* search algorithm has found the shortest path, the application visualizes the path by coloring the cells along the path blue.

Reset: Users can reset the grid to its initial state by pressing the C key.

User interface: The application has a simple and intuitive user interface that allows users to easily draw the grid and interact with the various features.